// src/api/orval/model/getApiOpenVpnClientsOverviewSeriesParams.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */
import type { OverviewGrouping } from "./overviewGrouping";

export type GetApiOpenVpnClientsOverviewSeriesParams = {
  From: string;
  To: string;
  Grouping?: OverviewGrouping;
  VpnServerId?: number;
  ExternalId?: string;
};