// src/api/orval/model/totalsPayloadDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */

export interface TotalsPayloadDto {
  sessionsCount?: number;
  usersCount?: number;
  trafficInBytes?: number;
  trafficOutBytes?: number;
  readonly trafficTotalBytes?: number;
}